requirements.txt
pypubsub>=4.0
adafruit-circuitpython-rgb-display
Pillow>=8.0
